Move everything in libpod into a single package for simplicity
