+++ libpod/container.go
@@ -1,4 +1,4 @@
-package ctr
+package libpod
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 
 var (
 	// ErrNotImplemented indicates that functionality is not yet implemented
-	ErrNotImplemented = fmt.Errorf("NOT IMPLEMENTED")
+	errNotImplemented = fmt.Errorf("NOT IMPLEMENTED")
 )
 
 // Container is a single OCI container
@@ -18,52 +18,52 @@ type Container struct {
 
 // Create creates a container in the OCI runtime
 func (c *Container) Create() error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Start starts a container
 func (c *Container) Start() error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Stop stops a container
 func (c *Container) Stop() error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Kill sends a signal to a container
 func (c *Container) Kill(signal uint) error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Exec starts a new process inside the container
 // Returns fully qualified URL of streaming server for executed process
 func (c *Container) Exec(cmd []string, tty bool, stdin bool) (string, error) {
-	return "", ErrNotImplemented
+	return "", errNotImplemented
 }
 
 // Attach attaches to a container
 // Returns fully qualified URL of streaming server for the container
 func (c *Container) Attach(stdin, tty bool) (string, error) {
-	return "", ErrNotImplemented
+	return "", errNotImplemented
 }
 
 // Mount mounts a container's filesystem on the host
 // The path where the container has been mounted is returned
 func (c *Container) Mount() (string, error) {
-	return "", ErrNotImplemented
+	return "", errNotImplemented
 }
 
 // Status gets a container's status
 // TODO this should return relevant information about container state
 func (c *Container) Status() error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Export exports a container's root filesystem as a tar archive
 // The archive will be saved as a file at the given path
 func (c *Container) Export(path string) error {
-	return ErrNotImplemented
+	return errNotImplemented
 }
 
 // Commit commits the changes between a container and its image, creating a new
@@ -73,5 +73,5 @@ func (c *Container) Export(path string) error {
 // a new base image will be created from the contents of the container's
 // filesystem
 func (c *Container) Commit() (*storage.Image, error) {
-	return nil, ErrNotImplemented
+	return nil, errNotImplemented
 }
